requirements.txt
requests==2.21.0
firebase_admin==3.1.0
pytz==2019.2
beautifulsoup4==4.8.2
protobuf==3.11.3
pandas==0.25.2
google-cloud-storage==1.26.0